move agents to agent subdirectory
